+++ docs/src/examples/components/Dropdown/Types/DropdownExample.shorthand.steps.ts
@@ -6,6 +6,7 @@ const selectors = {
 }
 
 const config: ScreenerTestsConfig = {
+  themes: ['teams', 'teamsDark', 'teamsHighContrast'],
   steps: [
     builder => builder.click(selectors.triggerButton).snapshot('Shows list'),
     builder =>